feat(zmq): make message access lazy
