+++ tests/integration/mime/test_segmenter.py
@@ -25,12 +25,12 @@ def test_dummy_seg():
 
 
 def test_dummy_seg_random():
-    f = Flow().add(uses=os.path.join(cur_dir, '../../yaml/dummy-seg-random.yml'))
+    f = Flow().add(uses=os.path.join(cur_dir, '../../unit/yaml/dummy-seg-random.yml'))
     with f:
         f.index(input_fn=random_docs(10, chunks_per_doc=0), output_fn=validate)
 
 
 def test_dummy_seg_not_random():
-    f = Flow().add(uses=os.path.join(cur_dir, '../../yaml/dummy-seg-not-random.yml'))
+    f = Flow().add(uses=os.path.join(cur_dir, '../../unit/yaml/dummy-seg-not-random.yml'))
     with f:
         f.index(input_fn=random_docs(10, chunks_per_doc=0), output_fn=validate)
